extract model classes up
