+++ src/main/java/dev/streamx/puresight/eds/Content.java
@@ -1,0 +1,18 @@
+package dev.streamx.puresight.eds;
+
+public class Content {
+
+  public String bytes;
+
+  public Content() {
+
+  }
+
+  public Content(String bytes) {
+    this.bytes = bytes;
+  }
+
+  public String getBytes() {
+    return bytes;
+  }
+}
