+++ src/main/java/dev/streamx/puresight/eds/Page.java
@@ -1,0 +1,18 @@
+package dev.streamx.puresight.eds;
+
+public class Page {
+
+  public Content content;
+
+  public Page() {
+
+  }
+
+  public Page(Content content) {
+    this.content = content;
+  }
+
+  public Content getContent() {
+    return content;
+  }
+}
